blurit/requirements.txt
asgiref==3.2.10
Django==3.0.7
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
opencv-python==4.2.0.34
tensorflow==2.2.0